requirements.txt
requests==2.31.0
python-dotenv==1.0.1
feedparser==6.0.11
schedule==1.2.2
openai==1.14.0
convex==0.7.0
